backend/requirements.txt
fastapi[all]
uvicorn
httpx
sqlalchemy
psycopg2-binary  # PostgreSQL 연결 드라이버
pgvector  # pgvector 지원